Refactor: 테스트 클래스 이름 변경
- 통합 테스트 클래스 IntegrationTest -> IT
