+++ src/test/java/com/uranus/taskmanager/api/invitation/controller/InvitationControllerTest.java
@@ -15,7 +15,6 @@ import org.springframework.mock.web.MockHttpSession;
 import com.uranus.taskmanager.api.invitation.exception.InvitationNotFoundException;
 import com.uranus.taskmanager.api.invitation.presentation.dto.response.AcceptInvitationResponse;
 import com.uranus.taskmanager.api.security.session.SessionAttributes;
-import com.uranus.taskmanager.api.workspace.domain.Workspace;
 import com.uranus.taskmanager.fixture.entity.InvitationEntityFixture;
 import com.uranus.taskmanager.fixture.entity.MemberEntityFixture;
 import com.uranus.taskmanager.fixture.entity.WorkspaceEntityFixture;
@@ -39,12 +38,7 @@ class InvitationControllerTest extends ControllerTestHelper {
 	void test1() throws Exception {
 		// given
 		String code = "TESTCODE";
-		Workspace workspace = Workspace.builder()
-			.code(code)
-			.name("workspace1")
-			.description("description1")
-			.build();
-
+		
 		AcceptInvitationResponse response = AcceptInvitationResponse.builder()
 			.invitationId(1L)
 			.build();
