+++ src/test/java/com/uranus/taskmanager/api/workspacemember/service/query/MemberWorkspaceQueryServiceIT.java
@@ -1,4 +1,4 @@
-package com.uranus.taskmanager.api.workspace.service;
+package com.uranus.taskmanager.api.workspacemember.service.query;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -7,7 +7,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.data.domain.PageRequest;
@@ -19,15 +18,13 @@ import com.uranus.taskmanager.api.member.domain.Member;
 import com.uranus.taskmanager.api.member.presentation.dto.request.SignupMemberRequest;
 import com.uranus.taskmanager.api.security.authentication.resolver.LoginMember;
 import com.uranus.taskmanager.api.workspace.domain.Workspace;
-import com.uranus.taskmanager.api.workspace.exception.WorkspaceNotFoundException;
 import com.uranus.taskmanager.api.workspace.presentation.dto.WorkspaceDetail;
 import com.uranus.taskmanager.api.workspacemember.WorkspaceRole;
-import com.uranus.taskmanager.api.workspacemember.exception.MemberNotInWorkspaceException;
 import com.uranus.taskmanager.api.workspacemember.presentation.dto.request.JoinWorkspaceRequest;
 import com.uranus.taskmanager.api.workspacemember.presentation.dto.response.MyWorkspacesResponse;
 import com.uranus.taskmanager.helper.ServiceIntegrationTestHelper;
 
-class WorkspaceQueryServiceTest extends ServiceIntegrationTestHelper {
+class MemberWorkspaceQueryServiceIT extends ServiceIntegrationTestHelper {
 
 	@BeforeEach
 	void setup() {
@@ -99,72 +96,6 @@ class WorkspaceQueryServiceTest extends ServiceIntegrationTestHelper {
 
 		// then
 		assertThat(response.getTotalElements()).isEqualTo(1);
-	}
-
-	@Transactional
-	@Test
-	@DisplayName("해당 워크스페이스에 참여하고 있으면, 워크스페이스의 코드로 상세 정보를 조회할 수 있다")
-	void test3() {
-		// given
-		memberService.signup(SignupMemberRequest.builder()
-			.loginId("member2")
-			.email("member2@test.com")
-			.password("member2password!")
-			.build());
-
-		LoginMember loginMember2 = LoginMember.builder()
-			.id(2L)
-			.loginId("member2")
-			.email("member2@test.com")
-			.build();
-
-		memberWorkspaceCommandService.joinWorkspace("TEST1111", new JoinWorkspaceRequest(), loginMember2.getId());
-
-		// when
-		WorkspaceDetail response = workspaceQueryService.getWorkspaceDetail("TEST1111");
-
-		// then
-		assertThat(response.getCode()).isEqualTo("TEST1111");
-		assertThat(response.getName()).isEqualTo("workspace1");
-	}
-
-	@Transactional
-	@Disabled("컨트롤러에서 해당 워크스페이스에서 VIEWER 이상만 호출가능하도록 변경.(쉽게 말해서 해당 워크스페이스에서 참여하는지 검증)")
-	@Test
-	@DisplayName("해당 워크스페이스에 참여하지 않으면, 유효한 코드로 상세 정보를 조회해도 예외가 발생한다")
-	void test4() {
-		// given
-		memberService.signup(SignupMemberRequest.builder()
-			.loginId("member2")
-			.email("member2@test.com")
-			.password("member2password!")
-			.build());
-
-		// LoginMember loginMember2 = LoginMember.builder()
-		// 	.id(2L)
-		// 	.loginId("member2")
-		// 	.email("member2@test.com")
-		// 	.build();
-
-		// when & then
-		assertThatThrownBy(() -> workspaceQueryService.getWorkspaceDetail("TEST1111"))
-			.isInstanceOf(MemberNotInWorkspaceException.class);
-	}
-
-	@Transactional
-	@Test
-	@DisplayName("유효하지 않은 코드로 워크스페이스를 조회하면 예외가 발생한다")
-	void test5() {
-		// given
-		memberService.signup(SignupMemberRequest.builder()
-			.loginId("member2")
-			.email("member2@test.com")
-			.password("member2password!")
-			.build());
-
-		// when & then
-		assertThatThrownBy(() -> workspaceQueryService.getWorkspaceDetail("BADCODE1"))
-			.isInstanceOf(WorkspaceNotFoundException.class);
 	}
 
 	@Transactional
