+++ src/test/java/com/uranus/taskmanager/api/workspacemember/service/query/WorkspaceMemberQueryIT.java
@@ -1,0 +1,4 @@
+package com.uranus.taskmanager.api.workspacemember.service.query;
+
+public class WorkspaceMemberQueryIT {
+}
