+++ src/test/java/com/uranus/taskmanager/helper/ControllerIntegrationTestHelper.java
@@ -1,0 +1,4 @@
+package com.uranus.taskmanager.helper;
+
+public class ControllerIntegrationTestHelper {
+}
