+++ src/test/java/com/uranus/taskmanager/helper/ServiceIntegrationTestHelper.java
@@ -18,16 +18,16 @@ import com.uranus.taskmanager.api.workspacemember.service.command.MemberWorkspac
 import com.uranus.taskmanager.api.workspacemember.service.command.WorkspaceMemberCommandService;
 import com.uranus.taskmanager.api.workspacemember.service.command.WorkspaceMemberInviteService;
 import com.uranus.taskmanager.api.workspacemember.service.query.MemberWorkspaceQueryService;
+import com.uranus.taskmanager.api.workspacemember.service.query.WorkspaceMemberQueryService;
 import com.uranus.taskmanager.fixture.repository.MemberRepositoryFixture;
 import com.uranus.taskmanager.fixture.repository.WorkspaceRepositoryFixture;
 import com.uranus.taskmanager.util.DatabaseCleaner;
 
 import jakarta.persistence.EntityManager;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 @SpringBootTest
 public abstract class ServiceIntegrationTestHelper {
+
 	/**
 	 * Common
 	 */
@@ -45,8 +45,8 @@ public abstract class ServiceIntegrationTestHelper {
 	protected WorkspaceMemberCommandService workspaceMemberCommandService;
 	@Autowired
 	protected WorkspaceMemberInviteService workspaceMemberInviteService;
-	// @Autowired
-	// protected WorkspaceMembershipQueryService workspaceMembershipQueryService;
+	@Autowired
+	protected WorkspaceMemberQueryService workspaceMemberQueryService;
 	@Autowired
 	protected MemberWorkspaceQueryService memberWorkspaceQueryService;
 	@Autowired
